+++ sdk/monitor/azquery/example_test.go
@@ -54,15 +54,6 @@ func ExampleNewMetricsClient() {
 	_ = client
 }
 
-func ExampleNewISO8601TimeInterval() {
-	// Use time.UTC
-	timespan, err := azquery.NewISO8601TimeInterval(time.Date(2022, 12, 25, 0, 0, 0, 0, time.UTC), time.Date(2022, 12, 25, 12, 0, 0, 0, time.UTC))
-	if err != nil {
-		//TODO: handle error
-	}
-	_ = timespan
-}
-
 func ExampleLogsClient_QueryWorkspace() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -73,17 +64,13 @@ func ExampleLogsClient_QueryWorkspace() {
 		//TODO: handle error
 	}
 	workspaceID := "g4d1e129-fb1e-4b0a-b234-250abc987ea65" // example Azure Log Analytics Workspace ID
-	timespan, err := azquery.NewISO8601TimeInterval(time.Date(2022, 12, 25, 0, 0, 0, 0, time.UTC), time.Date(2022, 12, 25, 12, 0, 0, 0, time.UTC))
-	if err != nil {
-		//TODO: handle error
-	}
 
 	res, err := client.QueryWorkspace(
 		context.TODO(),
 		workspaceID,
 		azquery.Body{
-			Query:    to.Ptr("AzureActivity | top 10 by TimeGenerated"),
-			Timespan: timespan,
+			Query:    to.Ptr("AzureActivity | top 10 by TimeGenerated"), // example Kusto query
+			Timespan: to.Ptr(azquery.NewISO8601TimeInterval(time.Date(2022, 12, 25, 0, 0, 0, 0, time.UTC), time.Date(2022, 12, 25, 12, 0, 0, 0, time.UTC))),
 		},
 		nil)
 	if err != nil {
@@ -112,22 +99,16 @@ func ExampleLogsClient_QueryWorkspace_second() {
 	}
 	workspaceID1 := "g4d1e129-fb1e-4b0a-b234-250abc987ea65" // example Azure Log Analytics Workspace ID
 	workspaceID2 := "h4bc4471-2e8c-4b1c-8f47-12b9a4d5ac71"
-	query := "search * | take 5"               // Example Kusto query
 	preferOptions := "include-statistics=true" // Advanced option to include stats. See docs: https://pkg.go.dev/github.com/Azure/azure-sdk-for-go/sdk/monitor/azquery#readme-increase-wait-time-include-statistics-include-render-visualization
-	timespan, err := azquery.NewISO8601TimeInterval(time.Date(2022, 12, 25, 0, 0, 0, 0, time.UTC), time.Date(2022, 12, 25, 12, 0, 0, 0, time.UTC))
-	if err != nil {
-		//TODO: handle error
-	}
-
 	res, err := client.QueryWorkspace(
 		context.TODO(),
 		workspaceID1,
 		azquery.Body{
-			Query:                &query,
-			Timespan:             timespan,
-			AdditionalWorkspaces: []*string{&workspaceID2},
+			Query:                to.Ptr(query),
+			Timespan:             to.Ptr(azquery.NewISO8601TimeInterval(time.Date(2022, 12, 25, 0, 0, 0, 0, time.UTC), time.Date(2022, 12, 25, 12, 0, 0, 0, time.UTC))),
+			AdditionalWorkspaces: []*string{to.Ptr(workspaceID2)},
 		},
-		&azquery.LogsClientQueryWorkspaceOptions{Prefer: &preferOptions})
+		to.Ptr(azquery.LogsClientQueryWorkspaceOptions{Prefer: to.Ptr(preferOptions)}))
 	if err != nil {
 		//TODO: handle error
 	}
@@ -163,15 +144,12 @@ func ExampleLogsClient_QueryBatch() {
 	}
 
 	workspaceID := "g4d1e129-fb1e-4b0a-b234-250abc987ea65" // example Azure Log Analytics Workspace ID
-	timespan, err := azquery.NewISO8601TimeInterval(time.Date(2022, 12, 25, 0, 0, 0, 0, time.UTC), time.Date(2022, 12, 25, 12, 0, 0, 0, time.UTC))
-	if err != nil {
-		//TODO: handle error
-	}
+	timespan := azquery.NewISO8601TimeInterval(time.Date(2022, 12, 25, 0, 0, 0, 0, time.UTC), time.Date(2022, 12, 25, 12, 0, 0, 0, time.UTC))
 
 	batchRequest := azquery.BatchRequest{[]*azquery.BatchQueryRequest{
-		{Body: &azquery.Body{Query: to.Ptr(kustoQuery1), Timespan: timespan}, CorrelationID: to.Ptr("1"), WorkspaceID: to.Ptr(workspaceID)},
-		{Body: &azquery.Body{Query: to.Ptr(kustoQuery2), Timespan: timespan}, CorrelationID: to.Ptr("2"), WorkspaceID: to.Ptr(workspaceID)},
-		{Body: &azquery.Body{Query: to.Ptr(kustoQuery3), Timespan: timespan}, CorrelationID: to.Ptr("3"), WorkspaceID: to.Ptr(workspaceID)},
+		{Body: &azquery.Body{Query: to.Ptr(kustoQuery1), Timespan: to.Ptr(timespan)}, CorrelationID: to.Ptr("1"), WorkspaceID: to.Ptr(workspaceID)},
+		{Body: &azquery.Body{Query: to.Ptr(kustoQuery2), Timespan: to.Ptr(timespan)}, CorrelationID: to.Ptr("2"), WorkspaceID: to.Ptr(workspaceID)},
+		{Body: &azquery.Body{Query: to.Ptr(kustoQuery3), Timespan: to.Ptr(timespan)}, CorrelationID: to.Ptr("3"), WorkspaceID: to.Ptr(workspaceID)},
 	}}
 
 	res, err := client.QueryBatch(context.TODO(), batchRequest, nil)
@@ -193,8 +171,13 @@ func ExampleMetricsClient_QueryResource() {
 	if err != nil {
 		//TODO: handle error
 	}
+
+	if err != nil {
+		//TODO: handle error
+	}
 	res, err := client.QueryResource(context.TODO(), resourceURI,
-		&azquery.MetricsClientQueryResourceOptions{Timespan: to.Ptr("2017-04-14T02:20:00Z/2017-04-14T04:20:00Z"),
+		&azquery.MetricsClientQueryResourceOptions{
+			Timespan:        to.Ptr(azquery.NewISO8601TimeInterval(time.Date(2022, 12, 25, 0, 0, 0, 0, time.UTC), time.Date(2022, 12, 25, 12, 0, 0, 0, time.UTC))),
 			Interval:        to.Ptr("PT1M"),
 			Metricnames:     nil,
 			Aggregation:     to.Ptr("Average,count"),