+++ sdk/monitor/azquery/CHANGELOG.md
@@ -3,10 +3,10 @@
 ## 0.4.0 (Unreleased)
 
 ### Features Added
-* Added ISO8601TimeInterval type with constructor to aide with logs timespan creation
+* Added ISO8601TimeInterval type with helper methods to aide with timespan creation
 
 ### Breaking Changes
-* Changed type of `Body.Timespan` from *string to *ISO8601TimeInterval
+* Changed type of `Body.Timespan`, `MetricsClientQueryResourceOptions.Timespan`, `Response.Timespan` from *string to *ISO8601TimeInterval
 * Remove `ColumnIndexLookup` field from Table struct
 * Renamed `Body.Workspaces` to `Body.AdditionalWorkspaces`
 * Renamed `Results.Render` and `BatchResponse.Render` to `Results.Visualization` and `BatchResponse.Visualization`